requirements.txt
google-adk>=1.14.0
mcp[cli]
python-dotenv
httpx
pydantic
litellm